Move aesara.compile.profiling.ScanProfileStats to aesara.scan.utils
