+++ aesara/scan/utils.py
@@ -11,6 +11,7 @@ import numpy as np
 
 from aesara import scalar as aes
 from aesara import tensor as at
+from aesara.compile.profiling import ProfileStats
 from aesara.configdefaults import config
 from aesara.graph.basic import (
     Constant,
@@ -121,6 +122,65 @@ class until:
     def __init__(self, condition):
         self.condition = at.as_tensor_variable(condition)
         assert self.condition.ndim == 0
+
+
+class ScanProfileStats(ProfileStats):
+    show_sum = False
+    callcount = 0.0
+    nbsteps = 0.0
+    call_time = 0.0
+
+    def __init__(self, atexit_print=True, name=None, **kwargs):
+        super().__init__(atexit_print, **kwargs)
+        self.name = name
+
+    def summary_globals(self, file):
+        # Do nothing, we don't want to print extra global summary
+        # here.
+        pass
+
+    def summary_function(self, file):
+        # RP: every time we compile a function a ProfileStats is created for
+        # that function. This means that every time a optimization replaces
+        # some scan op, some orphane ProfileStats remains in the air ..
+        # also even without any optimization, scan compiles a dummy function
+        # that will produce a ProfileStats that will correspond to a
+        # function that will never be called. Printing several empty
+        # Function profiling is just extremely confusing
+        if self.callcount == 0:
+            return
+        print("", file=file)
+
+        if self.name is not None:
+            print("Scan Op profiling (", self.name, ")", file=file)
+        else:
+            print("Scan Op profiling", file=file)
+        print("==================", file=file)
+        print(f"  Message: {self.message}", file=file)
+
+        print(
+            (
+                f"  Time in {self.callcount} calls of the op (for a total of {self.nbsteps} "
+                f"steps) {self.call_time:3}s"
+            ),
+            file=file,
+        )
+        print("", file=file)
+        val = 0
+        if self.call_time > 0:
+            val = self.vm_call_time * 100 / self.call_time
+        print(
+            f"  Total time spent in calling the VM {self.vm_call_time:e}s ({val:.3f}%)",
+            file=file,
+        )
+        val = 100
+        if self.call_time > 0:
+            val = 100.0 - self.vm_call_time * 100 / self.call_time
+        print(
+            f"  Total overhead (computing slices..) {self.call_time - self.vm_call_time:e}s ({val:.3f}%)",
+            file=file,
+        )
+        print("", file=file)
 
 
 def traverse(out, x, x_copy, d, visited=None):
